pyproject.toml
[project]
name = "lush-moments-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.119.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "passlib[bcrypt]>=1.7.0",
    "typer>=0.15.1",
    "rich>=13.9.4",
    "aiosmtplib>=3.0.2",
    "email-validator>=2.2.0",
    "httpx>=0.27.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.9",
    "langchain-google-genai>=3.0.0",
    "langchain>=1.0.2",
    "langgraph>=1.0.1",
]

[project.scripts]
lush-cli = "cli:app"
